agents/network-monitor/requirements.txt
# Core dependencies
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
ping3==4.0.4
httpx==0.25.2
asyncio==3.4.3

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0